control_residuos/requirements/base.txt
# Herramientas de construcción básicas
setuptools==80.9.0
wheel==0.45.1

# Web Framework
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1  # Actualizado para Python 3.13
Werkzeug==2.3.7
gunicorn==21.2.0
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0
colorama==0.4.6
Jinja2==3.1.6
MarkupSafe==3.0.3

# Database
SQLAlchemy==2.0.44  # Actualizado para Python 3.13
psycopg2-binary==2.9.11
alembic==1.12.0
greenlet==3.2.4  # Requerido por SQLAlchemy

# Utilidades del Sistema
python-dotenv==1.0.0
filelock==3.13.1
pathlib==1.0.1
PyYAML==6.0.1
python-dateutil==2.9.0.post0
six==1.17.0
packaging==23.2
typing-extensions==4.15.0
requests==2.32.5
charset-normalizer==3.4.4
urllib3==2.5.0
certifi==2025.10.5
idna==3.11